cr suggestions
